Update README.md
added some info on the 'affected' field. But I'm not sure about the example in the text reffering to 'VSC storage domain' as we do not have this as a specific category.
I also moved development below the semantics, as it is following better on the example in "how to use it", but feel free to ignore this.